requirements.txt

streamlit==1.43.2
python-docx
wxPython==4.2.1
pyperclip
openpyxl
